+++ src/pages/programsWip/Form.spec.tsx
@@ -1,0 +1,7 @@
+xdescribe('Program form tests', () => {
+    describe('Common', () => {
+        it('should not submit when required values are missing ', () => {})
+    })
+    describe('New', () => {})
+    describe('Edit', () => {})
+})